requirements.txt
pyodbc==4.0.30
python-dotenv~=0.13.0

Brotli==1.0.7
click==7.1.2
dash==1.14.0
dash-bootstrap-components==0.10.2
dash-core-components==1.10.2
dash-html-components==1.0.3
dash-renderer==1.6.0
dash-table==4.9.0
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.0
pandas==1.0.5
plotly==4.8.1
python-dateutil==2.8.1
pytz==2020.1
retrying==1.3.3
six==1.15.0
treelib==1.6.1
Werkzeug==1.0.1
xlrd==1.2.0